+++ src/vote/domain/candidate.ts
