+++ src/vote/domain/vote.status.ts
