+++ src/vote/domain/vote.ts
