+++ src/vote/domain/voting.log.ts
